подключение к БД и фикс возраста

+++ app/src/main/java/ru/wwerlosh/footprint/FirstBlockFragment.kt
@@ -81,7 +81,7 @@ class FirstBlockFragment : Fragment() {
             }
             GlobalData.name = name
             GlobalData.town = town
-            GlobalData.age = age
+            GlobalData.age = age.toInt()
             GlobalData.sex = sex
             val secondBlockFragment = SecondBlockFragment()
             val fragmentManager = requireActivity().supportFragmentManager
